# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getCountriesController","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"post":{"tags":["country-controller"],"operationId":"createCountryController","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestPost"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{languaje}/languaje":{"get":{"tags":["country-controller"],"operationId":"getContinentController","parameters":[],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getMostBordersController","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}}},"components":{"schemas":{"RequestPost":{"type":"object","properties":{"amountOfCountryToSave":{"type":"integer","format":"int32"}}},"CountryDTO":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}}}}}