# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CountryDTO {
    String code;
    String name;
}